Add Security Level and the four key states to RkpPoolStats
Attestation pool status is meaningful when it is associated with the
security level of the keymint instance. Counts of the keys in the four
different states are added as fields, by removing the enum.

Ignore-AOSP-First: No mergepath to AOSP.
Bug: 184301651
Test: Statsd test drive script.
Merged-In: I33264d937334cc48b65d5cc59f0522a3a20e512a

Change-Id: I33264d937334cc48b65d5cc59f0522a3a20e512a