Add support for slice_by_state to LocationTimeZoneProviderStateChanged
Add support for slice_by_state to LocationTimeZoneProviderStateChanged
for future releases (T+). This was suggested as a possible fix for
bug 195934779, but it cannot be used in Android S.

Documentation is not currently available, but it is assumed to be a
no-op until it is needed / used.

For reviewers:

The intent here is that each of the primary/secondary providers (as
indicated by provider_index) can only be in one state (as indicated by
state) at a time. The two providers change their states independently.

These annotations were suggested as the correct way to express this and
later enable slice_by_state so that the metrics system keeps track of
entry/exit of states for each provider (allowing for duration metrics
that track time each provider spends in each state).

Bug: 195934779
Test: Build only
Change-Id: I2ee0d31d28db664dad50e937c5c3da4fb4327928