Update the value of HotwordDetectorEvents atom
The first value in an enum must map to zero. Mapping the
first value to zero ensures the default behavior is consistent
between proto2 and proto3.

We add a unknown item to meet the first value to zero
instead of changing the index to avoid affecting the
original value.

Android Metrics Design Review : eldar/276723226

Bug: 207717787
Change-Id: Icd4a685043217834935bb6ca464200e4712cb784
Test: build pass