Improve the documentation of NeuralNetworks device_id fields
This CL improves the documentation of NeuralNetworks device_id fields,
clarifying that the device_id identifies which device drivers are
involved in the NeuralNetwork compilation or execution.

This CL also fixes the import order of neuralnetworks/enums.proto,
correcting it to be in alphabetical order.

Bug: 218915704
Test: mma
Change-Id: Ief6ca20c1a5320820eaaeafbe8e54c41ec3dc23a
Merged-In: Ief6ca20c1a5320820eaaeafbe8e54c41ec3dc23a