Remove a newline to make automerger happy
I accidentally removed a newline in a commit to tm-dev that exists
in tm-qpr-dev. This is causing merge conflicts when going from
tm-dev into tm-qpr-dev.

Remove the newline from the QPR branch so that it lines up with
tm-dev. This should fix up the automerger.

Test: n/a
Bug: 231495834
Change-Id: I1679222868af3d94fe90763a4c9f732d1ab18e11
Merged-In: I1679222868af3d94fe90763a4c9f732d1ab18e11